+++ pytorch_lightning/trainer/trainer.py
@@ -136,12 +136,12 @@ class Trainer(
         gradient_clip_algorithm: Optional[str] = None,
         process_position: int = 0,
         num_nodes: int = 1,
-        num_processes: int = 1,
+        num_processes: int = 1,  # TODO: Remove in 1.8
         devices: Optional[Union[List[int], str, int]] = None,
-        gpus: Optional[Union[List[int], str, int]] = None,
+        gpus: Optional[Union[List[int], str, int]] = None,  # TODO: Remove in 1.8
         auto_select_gpus: bool = False,
-        tpu_cores: Optional[Union[List[int], str, int]] = None,
-        ipus: Optional[int] = None,
+        tpu_cores: Optional[Union[List[int], str, int]] = None,  # TODO: Remove in 1.8
+        ipus: Optional[int] = None,  # TODO: Remove in 1.8
         log_gpu_memory: Optional[str] = None,  # TODO: Remove in 1.7
         progress_bar_refresh_rate: Optional[int] = None,  # TODO: remove in v1.7
         enable_progress_bar: bool = True,
@@ -262,6 +262,10 @@ class Trainer(
 
             gpus: Number of GPUs to train on (int) or which GPUs to train on (list or str) applied per node
 
+                .. deprecated:: v1.6
+                    ``gpus`` has been deprecated in v1.6 and will be removed in v1.8.
+                    Please use ``accelerator='gpu'`` and ``devices=x`` instead.
+
             gradient_clip_val: The value at which to clip gradients. Passing ``gradient_clip_val=None`` disables
                 gradient clipping. If using Automatic Mixed Precision (AMP), the gradients will be unscaled before.
 
@@ -349,6 +353,10 @@ class Trainer(
 
             num_processes: Number of processes for distributed training with ``accelerator="cpu"``.
 
+                .. deprecated:: v1.6
+                    ``num_processes`` has been deprecated in v1.6 and will be removed in v1.8.
+                    Please use ``accelerator='cpu'`` and ``devices=x`` instead.
+
             num_sanity_val_steps: Sanity check runs n validation batches before starting the training routine.
                 Set it to `-1` to run all batches in all validation dataloaders.
 
@@ -383,7 +391,15 @@ class Trainer(
 
             tpu_cores: How many TPU cores to train on (1 or 8) / Single TPU to train on [1]
 
+                .. deprecated:: v1.6
+                    ``tpu_cores`` has been deprecated in v1.6 and will be removed in v1.8.
+                    Please use ``accelerator='tpu'`` and ``devices=x`` instead.
+
             ipus: How many IPUs to train on.
+
+                .. deprecated:: v1.6
+                    ``ipus`` has been deprecated in v1.6 and will be removed in v1.8.
+                    Please use ``accelerator='ipu'`` and ``devices=x`` instead.
 
             track_grad_norm: -1 no tracking. Otherwise tracks that p-norm. May be set to 'inf' infinity-norm. If using
                 Automatic Mixed Precision (AMP), the gradients will be unscaled before logging them.