+++ tests/deprecated_api/test_remove_1-8.py
@@ -21,6 +21,7 @@ from pytorch_lightning.utilities.apply_func import move_data_to_device
 from pytorch_lightning.utilities.enums import DeviceType, DistributedType
 from pytorch_lightning.utilities.imports import _TORCHTEXT_LEGACY
 from tests.helpers.boring_model import BoringModel
+from tests.helpers.runif import RunIf
 from tests.helpers.torchtext_utils import get_dummy_torchtext_data_iterator
 
 
@@ -106,3 +107,26 @@ def test_v1_8_0_deprecated_on_hpc_hooks(tmpdir):
         match=r"Method `LightningModule.on_hpc_load` is deprecated in v1.6 and will be removed in v1.8."
     ):
         trainer.fit(load_model)
+
+
+def test_v1_8_0_deprecated_num_processes(tmpdir):
+    with pytest.deprecated_call(match=r"is deprecated in v1.6 and will be removed in v1.8."):
+        _ = Trainer(default_root_dir=tmpdir, num_processes=2)
+
+
+@RunIf(gpu=True)
+def test_v1_8_0_deprecated_gpus(tmpdir):
+    with pytest.deprecated_call(match=r"is deprecated in v1.6 and will be removed in v1.8."):
+        _ = Trainer(default_root_dir=tmpdir, gpus=2)
+
+
+@RunIf(tpu=True)
+def test_v1_8_0_deprecated_tpu_cores(tmpdir):
+    with pytest.deprecated_call(match=r"is deprecated in v1.6 and will be removed in v1.8."):
+        _ = Trainer(default_root_dir=tmpdir, tpu_cores=1)
+
+
+@RunIf(ipu=True)
+def test_v1_8_0_deprecated_ipus(tmpdir):
+    with pytest.deprecated_call(match=r"is deprecated in v1.6 and will be removed in v1.8."):
+        _ = Trainer(default_root_dir=tmpdir, ipus=2)